+++ packages/dataviz/src/components/RangeFilter/handlers/DateRangeHandler/DateRangeHandler.test.tsx
@@ -59,8 +59,39 @@ describe('DateRangeHandler', () => {
 		expect(max).toEqual(parseISO('2015-01-01T23:59:59.999').getTime());
 	});
 
-	it('Should format tick', () => {
-		const max = DateRangeHandler.tickFormatter!(new Date('2015-01-01T12:00:00').getTime());
-		expect(max).toEqual('2015-01-01');
+	it('Should create ticks', () => {
+		const limits = {
+			min: new Date('2000-01-01T12:00:00').getTime(),
+			max: new Date('2030-01-01T12:00:00').getTime(),
+		};
+		const ticks = DateRangeHandler.getTicks(limits);
+		expect(ticks).toEqual({
+			[limits.min]: '2000-01-01',
+			'1262300400000': '2010-01-01',
+			'1577833200000': '2020-01-01',
+			[limits.max]: '2030-01-01',
+		});
+	});
+	it('Should create ticks for tiny range', () => {
+		const limits = {
+			min: new Date('2000-01-01T12:00:00').getTime(),
+			max: new Date('2000-01-02T12:00:00').getTime(),
+		};
+		const ticks = DateRangeHandler.getTicks(limits);
+		expect(ticks).toEqual({
+			[limits.min]: '2000-01-01',
+			[limits.max]: '2000-01-02',
+		});
+	});
+	it('Should have min/max even for single day range', () => {
+		const limits = {
+			min: new Date('2020-01-01T00:00:00').getTime(),
+			max: new Date('2020-01-01T23:59:59').getTime(),
+		};
+		const ticks =  DateRangeHandler.getTicks(limits);
+		expect(ticks).toEqual({
+			[limits.min]: '2020-01-01',
+			[limits.max]: '2020-01-01',
+		});
 	});
 });